+++ src/main/java/com/grp2055/restbackend/repositories/BookingRepo.java
@@ -2,9 +2,9 @@ package com.grp2055.restbackend.repositories;
 
 import com.grp2055.restbackend.domain.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.RepositoryDefinition;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface BookingRepo extends JpaRepository<Booking, Long> {
+public interface BookingRepo extends JpaRepository<Booking, Integer> {
 }
