+++ src/main/java/com/grp2055/restbackend/repositories/UserRepo.java
@@ -4,8 +4,9 @@ import com.grp2055.restbackend.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface UserRepo extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, Integer> {
 
 
 }
