+++ src/main/java/com/grp2055/restbackend/service/RoomService.java
@@ -5,7 +5,7 @@ import com.grp2055.restbackend.domain.Room;
 import java.util.List;
 
 public interface RoomService  {
-    Room findRoomById (Long Id);
+    Room findRoomById (int Id);
     List<Room> findAllRooms();
     Room createNewRoom(Room room);
 }
